pyproject.toml
[tool.poetry]
name = "triplechartest"
version = "0.1.0"
description = "Test a char three times"
authors = ["Chris <dev@tgu-ltd.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.9.0.80"
pytest = "^8.1.1"
pyserial = "^3.5"
pyvisa = "^1.14.1"
pyvisa-py = "^0.7.2"
pytesseract = "^0.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
